Update mutations require authentication

+++ src/graphql-prisma/resolvers/Mutation.js
@@ -36,18 +36,22 @@ const Mutation = {
             token
         }
     },
-    updateUser(parent, {id, data}, {prisma}, info) {
+    updateUser(parent, {data}, {prisma, request}, info) {
+        const userId = getUserId(request);
+
         return prisma.mutation.updateUser({
             where: {
-                id
+                id: userId
             },
             data
         }, info);
     },
-    deleteUser(parent, {id}, {prisma}, info) {
+    deleteUser(parent, args, {prisma, request}, info) {
+        const userId = getUserId(request);
+
         return prisma.mutation.deleteUser({
             where: {
-                id
+                id: userId
             }
         });
     },
@@ -56,12 +60,36 @@ const Mutation = {
         data = {...data, author: {connect: {id: userId}}};
         return prisma.mutation.createPost({data}, info);
     },
-    deletePost(parent, {id}, {prisma}, info) {
+    async deletePost(parent, {id}, {prisma, request}, info) {
+        const userId = getUserId(request);
+        const postExists = await prisma.exists.Post({
+            id,
+            author: {
+                id: userId
+            }
+        });
+
+        if (!postExists) {
+            throw new Error('Unable to delete post')
+        }
+
         return prisma.mutation.deletePost({
             where: {id}
         }, info);
     },
-    updatePost(parent, {id, data}, {prisma}, info) {
+    async updatePost(parent, {id, data}, {prisma, request}, info) {
+        const userId = getUserId(request);
+        const postExists = await prisma.exists.Post({
+            id,
+            author: {
+                id: userId
+            }
+        });
+
+        if (!postExists) {
+            throw new Error('Unable to delete post')
+        }
+
         return prisma.mutation.updatePost({
             where: {
                 id
@@ -69,15 +97,29 @@ const Mutation = {
             data
         });
     },
-    createComment(parent, {data}, {prisma}, info) {
+    createComment(parent, {data}, {prisma, request}, info) {
+        const userId = getUserId(request);
+
         data = {
             ...data,
-            author: {connect: {id: data.author}},
+            author: {connect: {id: userId}},
             post: {connect: {id: data.post}},
         };
         return prisma.mutation.createComment({data}, info)
     },
-    updateComment(parent, {id, data}, {prisma}, info) {
+    async updateComment(parent, {id, data}, {prisma, request}, info) {
+        const userId = getUserId(request);
+        const commentExists = await prisma.exists.Comment({
+            id,
+            author: {
+                id: userId
+            }
+        });
+
+        if (!commentExists) {
+            throw new Error('Unable to update comment')
+        }
+
         return prisma.mutation.updateComment({
             where: {
                 id
@@ -85,7 +127,19 @@ const Mutation = {
             data
         }, info);
     },
-    deleteComment(parent, {id}, {prisma}, info) {
+    async deleteComment(parent, {id}, {prisma, request}, info) {
+        const userId = getUserId(request);
+        const commentExists = await prisma.exists.Comment({
+            id,
+            author: {
+                id: userId
+            }
+        });
+
+        if (!commentExists) {
+            throw new Error('Unable to delete comment')
+        }
+
         return prisma.mutation.deleteComment({
             where: {id}
         }, info);
